131 add api functions

+++ backend/db/migration/000011_create_conditions_table.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE conditions (
     climate_data_id INT NOT NULL,
     time_schedule_row_id INT NOT NULL,
     comparison_operator_id INT NOT NULL,
-    set_point INT NOT NULL,
+    set_point FLOAT NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (climate_data_id) REFERENCES climate_datas(id) ON DELETE RESTRICT,
